+++ src/__mocks__/get-secret.ts
@@ -2,7 +2,7 @@
 
 import awsSecretsManagerResponse from '../../test/fixtures/secrets/aws-secrets-manager-response.json';
 
-function getSecret(_: string, __: string): object {
+const getSecret = (_: string, __: string): object => {
   return JSON.parse(awsSecretsManagerResponse.Value);
 }