+++ test/secrets.test.ts
@@ -8,6 +8,9 @@ test('loading secrets from AWS Secrets Manager works', (): void => {
   expect(testConfig).toMatchObject({
     AWS_SECRETS_MANAGER_NAME: 'development/config-dug',
     DB_USERNAME: 'config-dug',
-    DB_PASSWORD: 'secret'
+    DB_PASSWORD: 'secret',
+    TEST_BOOLEAN: true,
+    TEST_INTEGER: 42,
+    TEST_FLOAT: 4.2
   });
 });