+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.2.0 (August 29, 2019)
+
+- Convert AWS Secrets Manager string values to correct types
+- Fix a bug with parsing integers and floats
+- Remove lodash
+
 ## 1.1.0 (June 14, 2019)
 
 Add support for environment specific local config files